[kernel] Fix decision whether to introduce an implicit cast or not

See merge request frama-c/frama-c!3654

+++ src/plugins/e-acsl/tests/arith/oracle/gen_integer_constant.c
@@ -12,7 +12,7 @@ int main(void)
 {
   int __retres;
   int x;
-  __e_acsl_memory_init((int *)0,(char ***)0,(size_t)8);
+  __e_acsl_memory_init((int *)0,(char ***)0,8UL);
   {
     __e_acsl_assert_data_t __gen_e_acsl_assert_data = {.values = (void *)0};
     __gen_e_acsl_assert_data.blocking = 1;
