+++ src/plugins/e-acsl/tests/bts/oracle/gen_issue-eacsl-145.c
@@ -14,7 +14,7 @@ void __e_acsl_globals_init(void)
   static char __e_acsl_already_run = 0;
   if (! __e_acsl_already_run) {
     __e_acsl_already_run = 1;
-    __e_acsl_store_block((void *)(& G),(size_t)4);
+    __e_acsl_store_block((void *)(& G),4UL);
     __e_acsl_full_init((void *)(& G));
   }
   return;
@@ -29,7 +29,7 @@ void __e_acsl_globals_clean(void)
 int main(void)
 {
   int __retres;
-  __e_acsl_memory_init((int *)0,(char ***)0,(size_t)8);
+  __e_acsl_memory_init((int *)0,(char ***)0,8UL);
   __e_acsl_globals_init();
   {
     int __gen_e_acsl_valid;
