+++ src/plugins/e-acsl/tests/builtin/oracle/gen_strcpy.c
@@ -105,9 +105,9 @@ void test_memory_tracking(void)
 {
   {
     char dest[4];
-    __e_acsl_store_block((void *)(dest),(size_t)4);
+    __e_acsl_store_block((void *)(dest),4UL);
     char src[2] = {(char)'b', (char)'\000'};
-    __e_acsl_store_block((void *)(src),(size_t)2);
+    __e_acsl_store_block((void *)(src),2UL);
     __e_acsl_full_init((void *)(& src));
     {
       int __gen_e_acsl_size;
@@ -262,9 +262,9 @@ void test_memory_tracking(void)
   }
   {
     char dest_0[4];
-    __e_acsl_store_block((void *)(dest_0),(size_t)4);
+    __e_acsl_store_block((void *)(dest_0),4UL);
     char src_0[4] = {(char)'a', (char)'b', (char)'\000'};
-    __e_acsl_store_block((void *)(src_0),(size_t)4);
+    __e_acsl_store_block((void *)(src_0),4UL);
     __e_acsl_full_init((void *)(& src_0));
     {
       int __gen_e_acsl_size_5;
@@ -342,7 +342,7 @@ void test_memory_tracking(void)
       __e_acsl_assert_clean(& __gen_e_acsl_assert_data_6);
     }
     /*@ assert \initialized(&src_0[0 .. 3]); */ ;
-    __e_acsl_builtin_strncpy(dest_0,(char const *)(src_0),(unsigned long)1);
+    __e_acsl_builtin_strncpy(dest_0,(char const *)(src_0),(size_t)1);
     {
       int __gen_e_acsl_initialized_7;
       __e_acsl_assert_data_t __gen_e_acsl_assert_data_7 =
@@ -410,9 +410,9 @@ void test_memory_tracking(void)
   }
   {
     char dest_1[4];
-    __e_acsl_store_block((void *)(dest_1),(size_t)4);
+    __e_acsl_store_block((void *)(dest_1),4UL);
     char src_1[4] = {(char)'b', (char)'\000'};
-    __e_acsl_store_block((void *)(src_1),(size_t)4);
+    __e_acsl_store_block((void *)(src_1),4UL);
     __e_acsl_full_init((void *)(& src_1));
     {
       int __gen_e_acsl_size_8;
@@ -490,7 +490,7 @@ void test_memory_tracking(void)
       __e_acsl_assert_clean(& __gen_e_acsl_assert_data_10);
     }
     /*@ assert \initialized(&src_1[0 .. 3]); */ ;
-    __e_acsl_builtin_strncpy(dest_1,(char const *)(src_1),(unsigned long)3);
+    __e_acsl_builtin_strncpy(dest_1,(char const *)(src_1),(size_t)3);
     {
       int __gen_e_acsl_size_10;
       int __gen_e_acsl_if_10;
@@ -589,12 +589,12 @@ char *__gen_e_acsl_strdup(char const *s)
 {
   __e_acsl_contract_t *__gen_e_acsl_contract;
   char *__retres;
-  __gen_e_acsl_contract = __e_acsl_contract_init((size_t)2);
+  __gen_e_acsl_contract = __e_acsl_contract_init(2UL);
   __retres = strdup(s);
   {
     int __gen_e_acsl_assumes_value;
     __gen_e_acsl_assumes_value = __e_acsl_contract_get_behavior_assumes
-    ((__e_acsl_contract_t const *)__gen_e_acsl_contract,(size_t)1);
+    ((__e_acsl_contract_t const *)__gen_e_acsl_contract,1UL);
     if (__gen_e_acsl_assumes_value) {
       __e_acsl_assert_data_t __gen_e_acsl_assert_data_4 =
         {.values = (void *)0};
@@ -680,15 +680,15 @@ pid_t __gen_e_acsl_waitpid(pid_t pid, int *stat_loc, int options)
   pid_t __retres;
   {
     int __gen_e_acsl_assumes_value;
-    __e_acsl_store_block((void *)(& stat_loc),(size_t)8);
+    __e_acsl_store_block((void *)(& stat_loc),8UL);
     __gen_e_acsl_at = stat_loc;
-    __gen_e_acsl_contract = __e_acsl_contract_init((size_t)2);
-    __e_acsl_contract_set_behavior_assumes(__gen_e_acsl_contract,(size_t)0,
+    __gen_e_acsl_contract = __e_acsl_contract_init(2UL);
+    __e_acsl_contract_set_behavior_assumes(__gen_e_acsl_contract,0UL,
                                            stat_loc == (int *)0);
-    __e_acsl_contract_set_behavior_assumes(__gen_e_acsl_contract,(size_t)1,
+    __e_acsl_contract_set_behavior_assumes(__gen_e_acsl_contract,1UL,
                                            stat_loc != (int *)0);
     __gen_e_acsl_assumes_value = __e_acsl_contract_get_behavior_assumes
-    ((__e_acsl_contract_t const *)__gen_e_acsl_contract,(size_t)1);
+    ((__e_acsl_contract_t const *)__gen_e_acsl_contract,1UL);
     if (__gen_e_acsl_assumes_value) {
       int __gen_e_acsl_valid;
       __e_acsl_assert_data_t __gen_e_acsl_assert_data =
@@ -929,18 +929,18 @@ void __e_acsl_globals_init(void)
 int main(int argc, char const **argv)
 {
   int __retres;
-  __e_acsl_memory_init(& argc,(char ***)(& argv),(size_t)8);
+  __e_acsl_memory_init(& argc,(char ***)(& argv),8UL);
   __e_acsl_globals_init();
   char empty_str[1] = {(char)'\000'};
   char *const_str = (char *)__gen_e_acsl_literal_string_6;
   char *src = __gen_e_acsl_strdup(__gen_e_acsl_literal_string_6);
-  char *dest1 = malloc((unsigned long)5);
-  char *dest2 = malloc((unsigned long)4);
+  char *dest1 = malloc((size_t)5);
+  char *dest2 = malloc((size_t)4);
   char dest3[256] =
     {(char)'a', (char)'b', (char)'c', (char)'d', (char)'\000'};
-  size_t len = (unsigned long)0;
-  char *unalloc_str = malloc((unsigned long)5);
-  char *_barrier = malloc((unsigned long)1);
+  size_t len = (size_t)0;
+  char *unalloc_str = malloc((size_t)5);
+  char *_barrier = malloc((size_t)1);
   free((void *)unalloc_str);
   {
     pid_t pid = __gen_e_acsl_fork();
@@ -950,7 +950,7 @@ int main(int argc, char const **argv)
     }
     else {
       int process_status;
-      __e_acsl_store_block((void *)(& process_status),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status),4UL);
       __gen_e_acsl_waitpid(pid,& process_status,0);
       signal_eval(process_status,0,__gen_e_acsl_literal_string_7);
       __e_acsl_delete_block((void *)(& process_status));
@@ -964,7 +964,7 @@ int main(int argc, char const **argv)
     }
     else {
       int process_status_0;
-      __e_acsl_store_block((void *)(& process_status_0),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_0),4UL);
       __gen_e_acsl_waitpid(pid_0,& process_status_0,0);
       signal_eval(process_status_0,0,__gen_e_acsl_literal_string_9);
       __e_acsl_delete_block((void *)(& process_status_0));
@@ -978,7 +978,7 @@ int main(int argc, char const **argv)
     }
     else {
       int process_status_1;
-      __e_acsl_store_block((void *)(& process_status_1),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_1),4UL);
       __gen_e_acsl_waitpid(pid_1,& process_status_1,0);
       signal_eval(process_status_1,1,__gen_e_acsl_literal_string_10);
       __e_acsl_delete_block((void *)(& process_status_1));
@@ -992,7 +992,7 @@ int main(int argc, char const **argv)
     }
     else {
       int process_status_2;
-      __e_acsl_store_block((void *)(& process_status_2),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_2),4UL);
       __gen_e_acsl_waitpid(pid_2,& process_status_2,0);
       signal_eval(process_status_2,1,__gen_e_acsl_literal_string_11);
       __e_acsl_delete_block((void *)(& process_status_2));
@@ -1006,7 +1006,7 @@ int main(int argc, char const **argv)
     }
     else {
       int process_status_3;
-      __e_acsl_store_block((void *)(& process_status_3),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_3),4UL);
       __gen_e_acsl_waitpid(pid_3,& process_status_3,0);
       signal_eval(process_status_3,1,__gen_e_acsl_literal_string_12);
       __e_acsl_delete_block((void *)(& process_status_3));
@@ -1020,7 +1020,7 @@ int main(int argc, char const **argv)
     }
     else {
       int process_status_4;
-      __e_acsl_store_block((void *)(& process_status_4),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_4),4UL);
       __gen_e_acsl_waitpid(pid_4,& process_status_4,0);
       signal_eval(process_status_4,0,__gen_e_acsl_literal_string_13);
       __e_acsl_delete_block((void *)(& process_status_4));
@@ -1034,7 +1034,7 @@ int main(int argc, char const **argv)
     }
     else {
       int process_status_5;
-      __e_acsl_store_block((void *)(& process_status_5),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_5),4UL);
       __gen_e_acsl_waitpid(pid_5,& process_status_5,0);
       signal_eval(process_status_5,1,__gen_e_acsl_literal_string_14);
       __e_acsl_delete_block((void *)(& process_status_5));
@@ -1048,7 +1048,7 @@ int main(int argc, char const **argv)
     }
     else {
       int process_status_6;
-      __e_acsl_store_block((void *)(& process_status_6),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_6),4UL);
       __gen_e_acsl_waitpid(pid_6,& process_status_6,0);
       signal_eval(process_status_6,0,__gen_e_acsl_literal_string_15);
       __e_acsl_delete_block((void *)(& process_status_6));
@@ -1062,7 +1062,7 @@ int main(int argc, char const **argv)
     }
     else {
       int process_status_7;
-      __e_acsl_store_block((void *)(& process_status_7),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_7),4UL);
       __gen_e_acsl_waitpid(pid_7,& process_status_7,0);
       signal_eval(process_status_7,1,__gen_e_acsl_literal_string_16);
       __e_acsl_delete_block((void *)(& process_status_7));
@@ -1071,12 +1071,12 @@ int main(int argc, char const **argv)
   {
     pid_t pid_8 = __gen_e_acsl_fork();
     if (! pid_8) {
-      __e_acsl_builtin_strncpy(dest1,(char const *)src,(unsigned long)5);
+      __e_acsl_builtin_strncpy(dest1,(char const *)src,(size_t)5);
       __gen_e_acsl_exit(0);
     }
     else {
       int process_status_8;
-      __e_acsl_store_block((void *)(& process_status_8),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_8),4UL);
       __gen_e_acsl_waitpid(pid_8,& process_status_8,0);
       signal_eval(process_status_8,0,__gen_e_acsl_literal_string_17);
       __e_acsl_delete_block((void *)(& process_status_8));
@@ -1085,12 +1085,12 @@ int main(int argc, char const **argv)
   {
     pid_t pid_9 = __gen_e_acsl_fork();
     if (! pid_9) {
-      __e_acsl_builtin_strncpy(dest1,(char const *)src,(unsigned long)6);
+      __e_acsl_builtin_strncpy(dest1,(char const *)src,(size_t)6);
       __gen_e_acsl_exit(0);
     }
     else {
       int process_status_9;
-      __e_acsl_store_block((void *)(& process_status_9),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_9),4UL);
       __gen_e_acsl_waitpid(pid_9,& process_status_9,0);
       signal_eval(process_status_9,1,__gen_e_acsl_literal_string_18);
       __e_acsl_delete_block((void *)(& process_status_9));
@@ -1099,13 +1099,12 @@ int main(int argc, char const **argv)
   {
     pid_t pid_10 = __gen_e_acsl_fork();
     if (! pid_10) {
-      __e_acsl_builtin_strncpy(unalloc_str,(char const *)src,
-                               (unsigned long)5);
+      __e_acsl_builtin_strncpy(unalloc_str,(char const *)src,(size_t)5);
       __gen_e_acsl_exit(0);
     }
     else {
       int process_status_10;
-      __e_acsl_store_block((void *)(& process_status_10),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_10),4UL);
       __gen_e_acsl_waitpid(pid_10,& process_status_10,0);
       signal_eval(process_status_10,1,__gen_e_acsl_literal_string_19);
       __e_acsl_delete_block((void *)(& process_status_10));
@@ -1114,12 +1113,12 @@ int main(int argc, char const **argv)
   {
     pid_t pid_11 = __gen_e_acsl_fork();
     if (! pid_11) {
-      __e_acsl_builtin_strncpy(const_str,(char const *)src,(unsigned long)5);
+      __e_acsl_builtin_strncpy(const_str,(char const *)src,(size_t)5);
       __gen_e_acsl_exit(0);
     }
     else {
       int process_status_11;
-      __e_acsl_store_block((void *)(& process_status_11),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_11),4UL);
       __gen_e_acsl_waitpid(pid_11,& process_status_11,0);
       signal_eval(process_status_11,1,__gen_e_acsl_literal_string_20);
       __e_acsl_delete_block((void *)(& process_status_11));
@@ -1128,12 +1127,12 @@ int main(int argc, char const **argv)
   {
     pid_t pid_12 = __gen_e_acsl_fork();
     if (! pid_12) {
-      __e_acsl_builtin_strncpy(src,(char const *)const_str,(unsigned long)5);
+      __e_acsl_builtin_strncpy(src,(char const *)const_str,(size_t)5);
       __gen_e_acsl_exit(0);
     }
     else {
       int process_status_12;
-      __e_acsl_store_block((void *)(& process_status_12),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_12),4UL);
       __gen_e_acsl_waitpid(pid_12,& process_status_12,0);
       signal_eval(process_status_12,0,__gen_e_acsl_literal_string_21);
       __e_acsl_delete_block((void *)(& process_status_12));
@@ -1142,12 +1141,12 @@ int main(int argc, char const **argv)
   {
     pid_t pid_13 = __gen_e_acsl_fork();
     if (! pid_13) {
-      __e_acsl_builtin_strncpy(src,(char const *)src,(unsigned long)5);
+      __e_acsl_builtin_strncpy(src,(char const *)src,(size_t)5);
       __gen_e_acsl_exit(0);
     }
     else {
       int process_status_13;
-      __e_acsl_store_block((void *)(& process_status_13),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_13),4UL);
       __gen_e_acsl_waitpid(pid_13,& process_status_13,0);
       signal_eval(process_status_13,1,__gen_e_acsl_literal_string_22);
       __e_acsl_delete_block((void *)(& process_status_13));
@@ -1156,13 +1155,12 @@ int main(int argc, char const **argv)
   {
     pid_t pid_14 = __gen_e_acsl_fork();
     if (! pid_14) {
-      __e_acsl_builtin_strncpy(& dest3[5],(char const *)(dest3),
-                               (unsigned long)5);
+      __e_acsl_builtin_strncpy(& dest3[5],(char const *)(dest3),(size_t)5);
       __gen_e_acsl_exit(0);
     }
     else {
       int process_status_14;
-      __e_acsl_store_block((void *)(& process_status_14),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_14),4UL);
       __gen_e_acsl_waitpid(pid_14,& process_status_14,0);
       signal_eval(process_status_14,0,__gen_e_acsl_literal_string_23);
       __e_acsl_delete_block((void *)(& process_status_14));
@@ -1171,13 +1169,12 @@ int main(int argc, char const **argv)
   {
     pid_t pid_15 = __gen_e_acsl_fork();
     if (! pid_15) {
-      __e_acsl_builtin_strncpy(& dest3[4],(char const *)(dest3),
-                               (unsigned long)5);
+      __e_acsl_builtin_strncpy(& dest3[4],(char const *)(dest3),(size_t)5);
       __gen_e_acsl_exit(0);
     }
     else {
       int process_status_15;
-      __e_acsl_store_block((void *)(& process_status_15),(size_t)4);
+      __e_acsl_store_block((void *)(& process_status_15),4UL);
       __gen_e_acsl_waitpid(pid_15,& process_status_15,0);
       signal_eval(process_status_15,1,__gen_e_acsl_literal_string_24);
       __e_acsl_delete_block((void *)(& process_status_15));
