+++ src/plugins/e-acsl/tests/constructs/oracle/gen_loop.c
@@ -403,7 +403,7 @@ void unnatural_loop(void)
 int main(void)
 {
   int __retres;
-  __e_acsl_memory_init((int *)0,(char ***)0,(size_t)8);
+  __e_acsl_memory_init((int *)0,(char ***)0,8UL);
   simple_loop();
   nested_loops();
   unnatural_loop();
